# apps/backend/swagger-output/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/whatsapp-agents/me": {
      "get": {
        "operationId": "WhatsAppAgentController_getMyAgent",
        "summary": "Get current user's WhatsApp agent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the WhatsApp agent for the current user"
          },
          "404": {
            "description": "WhatsApp agent not found for this user"
          }
        },
        "tags": ["whatsapp-agents"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/whatsapp-agents/provision": {
      "post": {
        "operationId": "WhatsAppAgentController_provisionAgent",
        "summary": "Provision a WhatsApp agent for the current user",
        "parameters": [],
        "responses": {
          "201": {
            "description": "WhatsApp agent successfully provisioned"
          },
          "409": {
            "description": "WhatsApp agent already exists for this user"
          }
        },
        "tags": ["whatsapp-agents"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/whatsapp-agents/{id}/status": {
      "patch": {
        "operationId": "WhatsAppAgentController_updateAgentStatus",
        "summary": "Update agent status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "WhatsApp agent ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent status successfully updated"
          },
          "404": {
            "description": "WhatsApp agent not found"
          }
        },
        "tags": ["whatsapp-agents"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/users/me": {
      "get": {
        "operationId": "UserController_getMe",
        "summary": "Get current user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": ["users"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UserController_updateMe",
        "summary": "Update current user profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": ["users"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/me/import-whatsapp": {
      "post": {
        "operationId": "UserController_importWhatsAppData",
        "summary": "Import WhatsApp Business data for current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "WhatsApp data imported successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportWhatsAppDataResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "WhatsApp agent not found for this user"
          },
          "500": {
            "description": "Failed to import WhatsApp data"
          }
        },
        "tags": ["users"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/me/stats": {
      "get": {
        "operationId": "UserController_getStats",
        "summary": "Get current user statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User statistics retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": ["users"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/webhooks/whatsapp/connected": {
      "post": {
        "operationId": "WebhooksController_whatsappConnected",
        "summary": "Webhook appelé quand WhatsApp se connecte",
        "description": "Endpoint appelé par le whatsapp-connector quand le client WhatsApp est prêt",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatsAppConnectedDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion traitée avec succès"
          }
        },
        "tags": ["webhooks"]
      }
    }
  },
  "info": {
    "title": "Monorepo API",
    "description": "Interractive documentation for the Monorepo API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UpdateAgentStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["PROVISIONING", "RUNNING", "STOPPED", "ERROR", "DELETED"],
            "description": "The new status of the agent",
            "example": "RUNNING"
          },
          "connectionStatus": {
            "type": "string",
            "enum": [
              "CONNECTED",
              "DISCONNECTED",
              "QR_REQUIRED",
              "PAIRING_REQUIRED"
            ],
            "description": "The connection status of WhatsApp",
            "example": "CONNECTED"
          }
        },
        "required": ["status"]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "phoneNumber": {
            "type": "string",
            "description": "User phone number"
          }
        }
      },
      "ImportWhatsAppDataResponseDto": {
        "type": "object",
        "properties": {
          "businessInfo": {
            "type": "object",
            "description": "Business information imported from WhatsApp"
          },
          "productsImported": {
            "type": "number",
            "description": "Number of products imported from catalog"
          },
          "contactsImported": {
            "type": "number",
            "description": "Number of contacts imported"
          }
        },
        "required": ["businessInfo", "productsImported", "contactsImported"]
      },
...
      "WhatsAppConnectedDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}